cleanup driverinfo handling
