+++ go/adbc/driver/driverbase/connection.go
@@ -20,6 +20,8 @@ package driverbase
 import (
 	"context"
 	"fmt"
+	"runtime/debug"
+	"strings"
 
 	"github.com/apache/arrow-adbc/go/adbc"
 	"github.com/apache/arrow-adbc/go/adbc/driver/internal"
@@ -34,6 +36,24 @@ const (
 	ConnectionMessageCannotCommit      = "Cannot commit when autocommit is enabled"
 	ConnectionMessageCannotRollback    = "Cannot rollback when autocommit is enabled"
 )
+
+var (
+	infoDriverVersion      string
+	infoDriverArrowVersion string
+)
+
+func init() {
+	if info, ok := debug.ReadBuildInfo(); ok {
+		for _, dep := range info.Deps {
+			switch {
+			case dep.Path == "github.com/apache/arrow-adbc/go/adbc":
+				infoDriverVersion = dep.Version
+			case strings.HasPrefix(dep.Path, "github.com/apache/arrow/go/"):
+				infoDriverArrowVersion = dep.Version
+			}
+		}
+	}
+}
 
 // ConnectionImpl is an interface that drivers implement to provide
 // vendor-specific functionality.
@@ -97,6 +117,17 @@ type ConnectionImplBase struct {
 //   - database is a DatabaseImplBase containing the common resources from the parent
 //     database, allowing the Arrow allocator, error handler, and logger to be reused.
 func NewConnectionImplBase(database *DatabaseImplBase) ConnectionImplBase {
+	if infoDriverVersion != "" {
+		if err := database.DriverInfo.RegisterInfoCode(adbc.InfoDriverVersion, infoDriverVersion); err != nil {
+			panic(err)
+		}
+	}
+	if infoDriverArrowVersion != "" {
+		if err := database.DriverInfo.RegisterInfoCode(adbc.InfoDriverArrowVersion, infoDriverArrowVersion); err != nil {
+			panic(err)
+		}
+	}
+
 	return ConnectionImplBase{
 		Alloc:       database.Alloc,
 		ErrorHelper: database.ErrorHelper,
